requirements.txt
nicegui~=1.4.36
pywebview~=5.2
qrcode~=7.4.2
requests~=2.32.3
httpx~=0.27.2